Add parallel workers
