+++ docker/runner/config/root
@@ -1,5 +1,4 @@
 10 1 */15 * * /usr/bin/python /home/domain_statistic/update_as_info.py >> /home/domain_statistic/download/update_as_info.log
-10 0 * * * /usr/bin/python /home/domain_statistic/update_domain.py -n `cat /etc/resolv.conf | awk '{print $2}'` >> /home/domain_statistic/download/update_domain.log
+10 0 * * * /usr/bin/python /home/domain_statistic/update_domain.py -n `cat /etc/resolv.conf | awk '{print $2}'` >> /home/domain_statistic/download/update_domain.log; /usr/bin/python /home/domain_statistic/update_statistic.py;
 30 12 * * * /usr/bin/python /home/domain_statistic/update_regru.py >> /home/domain_statistic/download/update_regru.log
-30 12 * * * /usr/bin/python /home/domain_statistic/update_statistic.py
 10 10 */30 * * /usr/bin/python /home/domain_statistic/normalization.py
